requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0

# OpenAI API
openai>=1.0.0
tiktoken>=0.5.0

# Optional: For enhanced evaluation metrics
# scikit-learn>=1.0.0

# Development dependencies
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
